requirements.txt
asgiref==3.8.1
attrs==24.2.0
Django==5.1
django-cors-headers==4.4.0
django-debug-toolbar==4.4.6
django-filter==24.3
django-resized==1.0.2
django-tinymce==4.1.0
django-unfold==0.37.0
djangorestframework==3.15.2
drf-spectacular==0.27.2
drf-spectacular-sidecar==2024.7.1
hashids==1.3.1
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
Markdown==3.7
pillow==10.4.0
python-dotenv==1.0.1
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.20.0
sqlparse==0.5.1
uritemplate==4.1.1
